requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Pillow==10.3.0
tensorflow==2.18.0
numpy==1.26.4